requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
